+++ src/cqportal/shieldwall/StraightBasic.py
@@ -1,6 +1,6 @@
 import cadquery as cq
 from cadqueryhelper import Base
-from . import ShieldShape
+from . import ShieldShape, Magnets
 
 class StraightBasic(Base):
     def __init__(self):
@@ -10,15 +10,26 @@ class StraightBasic(Base):
         self.length = 75
         self.width = 20
         self.height = 20
+        self.base_height = 5
+        self.magnet_padding = 1
+        self.render_magnets = True
 
         # blueprints
-        self.bp_shape = ShieldShape()
+        self.shape_bp = ShieldShape()
+        self.magnets_bp = Magnets()
 
         #shape
         self.wall = None
+        
+    def __make_shape(self, parent=None):
+        self.shape_bp.length = self.height
+        self.shape_bp.width = self.width
+        self.shape_bp.base_height = self.base_height
+        self.shape_bp.make(parent)
+        
 
     def __make_wall_basic(self):
-        shape = self.bp_shape.shape
+        shape = self.shape_bp.shape
         base_wall = (
             shape
             .extrude(self.length)
@@ -27,14 +38,32 @@ class StraightBasic(Base):
         )
         self.wall = base_wall
         
+    def __make_magnets(self):
+        self.magnets_bp.distance = self.width - self.magnets_bp.pip_radius*2 - self.magnet_padding*2
+        self.magnets_bp.make()
+        
     def make(self, parent=None):
         super().make(parent)
-        self.bp_shape.length = self.height
-        self.bp_shape.width = self.width
-        self.bp_shape.make(parent)
 
+        self.__make_shape(parent)
         self.__make_wall_basic()
+        self.__make_magnets()
         
     def build(self):
         super().build()
-        return self.wall
+        magnets = self.magnets_bp.build()
+        magnet_x = self.length/2 - self.magnets_bp.pip_height/2
+        magnet_z = -(self.height/2) + self.base_height - self.magnets_bp.pip_radius - self.magnet_padding
+        scene =(
+            cq.Workplane("XY")
+            .union(self.wall)
+
+        )
+        
+        if self.render_magnets:
+            scene = (
+                scene
+                .cut(magnets.translate((magnet_x,0,magnet_z)))
+                .cut(magnets.translate((-magnet_x,0,magnet_z)))
+            )
+        return scene